feat: implement login handler with authentication and error handling
