+++ nicknamer/server/src/web/mod.rs
@@ -1,7 +1,6 @@
 use askama::Template;
-use axum::extract::{Form, State};
-use axum::http::{HeaderMap, HeaderName, HeaderValue, StatusCode};
-use axum::response::{Html, IntoResponse, Response};
+use axum::http::StatusCode;
+use axum::response::Html;
 use migration::MigratorTrait;
 use sea_orm::Database;
 use std::sync::Arc;
@@ -9,6 +8,7 @@ use tower::ServiceBuilder;
 use tower_http::trace::{DefaultMakeSpan, DefaultOnRequest, DefaultOnResponse, TraceLayer};
 use tracing::Level;
 
+use crate::auth::{AppState, login_handler};
 use crate::config;
 use crate::web::middleware::cors_expose_headers;
 
@@ -36,18 +36,6 @@ impl axum::response::IntoResponse for WebError {
         )
             .into_response()
     }
-}
-
-#[derive(Clone)]
-struct AppState {
-    config: Arc<config::Config>,
-}
-
-/// Represents the login request payload.
-#[derive(serde::Deserialize, Debug)]
-struct LoginRequest {
-    username: String,
-    password: String,
 }
 
 #[tracing::instrument]
@@ -86,41 +74,6 @@ async fn health_check_handler() -> &'static str {
     "OK"
 }
 
-/// Handles the login request.
-/// Checks submitted username and password against admin credentials.
-async fn login_handler(
-    State(state): State<AppState>,
-    Form(payload): Form<LoginRequest>,
-) -> Result<Response, WebError> {
-    if payload.username == state.config.admin_username
-        && payload.password == state.config.admin_password
-    {
-        let html = LoginSuccessTemplate {
-            name: &payload.username,
-        }
-        .render()
-        .map_err(WebError::from)?;
-
-        Ok(Html(html).into_response())
-    } else {
-        let error_message = LoginErrorMessageTemplate.render().map_err(WebError::from)?;
-
-        let mut headers = HeaderMap::new();
-        headers.insert(
-            HeaderName::from_static("hx-retarget"),
-            HeaderValue::from_static("#login-message"),
-        );
-        headers.insert(
-            HeaderName::from_static("hx-reswap"),
-            HeaderValue::from_static("outerHTML"),
-        );
-
-        let mut response = Html(error_message).into_response();
-        response.headers_mut().extend(headers);
-        Ok(response)
-    }
-}
-
 async fn welcome_handler() -> Result<Html<String>, WebError> {
     IndexTemplate.render().map(Html).map_err(WebError::from)
 }
@@ -129,118 +82,16 @@ async fn welcome_handler() -> Result<Html<String>, WebError> {
 #[template(path = "index.html")]
 struct IndexTemplate;
 
-#[derive(Template)]
-#[template(path = "login_success.html")]
-struct LoginSuccessTemplate<'a> {
-    name: &'a str,
-}
-
-#[derive(Template)]
-#[template(path = "login_error_message.html")]
-struct LoginErrorMessageTemplate;
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::config::Config;
-    use axum::Router;
-    use axum::body::Body;
-    use axum::http::{Request, StatusCode};
-    use tower::ServiceExt; // for `oneshot`
-
-    async fn test_app(config: Config) -> axum::Router {
-        let state = AppState {
-            config: Arc::new(config),
-        };
-        Router::new()
-            .route("/login", axum::routing::post(login_handler))
-            .with_state(state)
-    }
+    use axum::http::StatusCode;
+    use axum::response::IntoResponse;
 
     #[tokio::test]
     async fn can_check_health() {
         let response = health_check_handler().await;
         assert_eq!(response, "OK");
-    }
-
-    #[tokio::test]
-    async fn can_login_with_valid_credentials() {
-        let config = Config {
-            db_url: "".to_string(),
-            port: 8080,
-            admin_username: "admin".to_string(),
-            admin_password: "password".to_string(),
-        };
-        let app = test_app(config).await;
-
-        let response = app
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri("/login")
-                    .header("content-type", "application/x-www-form-urlencoded")
-                    .body(Body::from("username=admin&password=password"))
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(response.status(), StatusCode::OK);
-
-        let body = axum::body::to_bytes(response.into_body(), usize::MAX)
-            .await
-            .unwrap();
-        assert_eq!(
-            body,
-            LoginSuccessTemplate { name: "admin" }.render().unwrap()
-        );
-    }
-
-    #[tokio::test]
-    async fn can_reject_invalid_credentials() {
-        let config = Config {
-            db_url: "".to_string(),
-            port: 8080,
-            admin_username: "admin".to_string(),
-            admin_password: "password".to_string(),
-        };
-        let app = test_app(config).await;
-
-        let response = app
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri("/login")
-                    .header("content-type", "application/x-www-form-urlencoded")
-                    .body(Body::from("username=wrong&password=wrong"))
-                    .unwrap(),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(response.status(), StatusCode::OK);
-
-        // Check HX-Retarget header
-        let hx_retarget = response.headers().get("hx-retarget");
-        assert_eq!(
-            hx_retarget,
-            Some(&axum::http::HeaderValue::from_static("#login-message"))
-        );
-
-        // Check HX-Reswap header
-        let hx_reswap = response.headers().get("hx-reswap");
-        assert_eq!(
-            hx_reswap,
-            Some(&axum::http::HeaderValue::from_static("outerHTML"))
-        );
-
-        let body = axum::body::to_bytes(response.into_body(), usize::MAX)
-            .await
-            .unwrap();
-        let rendered_error = LoginErrorMessageTemplate.render().unwrap();
-        assert_eq!(body, rendered_error);
-        // Verify the error message is included in the response
-        assert!(rendered_error.contains("Login failed. Please try again."));
     }
 
     #[tokio::test]
